fix: incorrect blockchain.relayfee description
The description of `blockchain.relayfee` promises the user that
transactions paying the relayfee will get accepted into the servers
mempool. This however seems incorrect as the relayfee is just a
fixed value in the core backend (either configured or a hardcoded default value)
and doesn't change with a growing mempool.
See definition:
https://github.com/laanwj/bitcoin/blob/d387b8ec15e553db9b9c370314ee359e981fb374/src/main.cpp#L55-L56
and usage for the API:
https://github.com/laanwj/bitcoin/blob/d387b8ec15e553db9b9c370314ee359e981fb374/src/rpcmisc.cpp#L86-L87

To get the minimal feerate to be paid for mempool acceptance one has to
use getmempoolinfo['mempoolminfee'].